requirement.txt
pypdf
scikit-learn
sentence-transformers
faiss-cpu
numpy
pandas
tqdm
streamlit
